init reportistica

+++ database/migrations/2021_05_16_201732_create_post_table.php
@@ -15,10 +15,10 @@
       public function up(): void {
          if(!Schema::hasTable('Post')) {
             Schema::create('Post', function (Blueprint $table) {
-               $table->increments('id')->comment('Identificativo Intero del Post dell\' Utente');
+               $table->bigIncrements('id')->comment('Identificativo Intero del Post dell\' Utente');
                $table->string('testo', 255)->comment('Testo del Post dell\' Utente');
                $table->string('foto', 25)->comment('Foto del Post dell\' Utente (relativa path del file)');
-               $table->unsignedInteger('utente_id')->comment('Riferimento alla Chiava Primaria di Utente');
+               $table->unsignedBigInteger('utente_id')->comment('Riferimento alla Chiava Primaria di Utente');
                $table->timestamp('created_at')->useCurrent()->comment('Data Creazione del Record (consente di ottenere anche la di Pubblicazione del Post)');
                $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP() ON UPDATE CURRENT_TIMESTAMP()'))->comment('Data Aggiornamento del Record');
                $table->unique(['utente_id', 'created_at', 'testo'], 'utenteTestoCreatedAt_UtentePost_UNIQUE');
