+++ database/migrations/2021_05_16_201732_create_utentelavoro_table.php
@@ -14,7 +14,7 @@
       public function up(): void {
          if(!Schema::hasTable('UtenteLavoro')) {
             Schema::create('UtenteLavoro', function (Blueprint $table) {
-               $table->unsignedInteger('utente_id')->comment('Riferimento alla Chiave Primaria di Utente');
+               $table->unsignedBigInteger('utente_id')->comment('Riferimento alla Chiave Primaria di Utente');
                $table->unsignedInteger('lavoro_id')->default(1)->index('LavoroUtenteFK')->comment('Riferimento alla Chiave Primaria di Lavoro');
                $table->date('dataInizioLavoro')->nullable()->comment('Data inizio Lavoro dell\' Utente');
                $table->primary(['utente_id', 'lavoro_id']);
